Updated the documentation for docker 1.9
In previous versions of docker, specifying a `-p` flag along with `--net=host` was ignored. After the networking updates in docker 1.9, these now conflict and docker will not start the container. This change is backwards compatible since the `-p` flag has always been ignored when specified with `--net=host`.

Fixes #43